[+] android clang type errors fixed

See merge request cellframe/cellframe-sdk!1386

+++ modules/type/blocks/dap_chain_cs_blocks.c
@@ -135,7 +135,7 @@ static void s_callback_cs_blocks_purge(dap_chain_t *a_chain);
 static dap_chain_block_t *s_new_block_move(dap_chain_cs_blocks_t *a_blocks, size_t *a_new_block_size);
 
 //Work with atoms
-static uint64_t s_callback_count_atom(dap_chain_t *a_chain);
+static size_t s_callback_count_atom(dap_chain_t *a_chain);
 static dap_list_t *s_callback_get_atoms(dap_chain_t *a_chain, size_t a_count, size_t a_page, bool a_reverse);
 
 static bool s_seed_mode = false;
@@ -1814,7 +1814,7 @@ static size_t s_callback_add_datums(dap_chain_t *a_chain, dap_chain_datum_t **a_
  * @param a_chain Chain object
  * @return size_t
  */
-static uint64_t s_callback_count_atom(dap_chain_t *a_chain)
+static size_t s_callback_count_atom(dap_chain_t *a_chain)
 {
     dap_chain_cs_blocks_t *l_blocks = DAP_CHAIN_CS_BLOCKS(a_chain);
     assert(l_blocks && l_blocks->chain == a_chain);
